Retarget AssertJIterableRules
