+++ error-prone-contrib/src/test/java/tech/picnic/errorprone/refasterrules/RefasterRulesTest.java
@@ -19,7 +19,6 @@ final class RefasterRulesTest {
           AssertJBooleanRules.class,
           AssertJByteRules.class,
           AssertJCharSequenceRules.class,
-          AssertJCollectionRules.class,
           AssertJComparableRules.class,
           AssertJDoubleRules.class,
           AssertJDurationRules.class,
